+++ TrainingProject/src/main/java/com/example/service/UserServiceImpl.java
@@ -1,37 +1,45 @@
 package com.example.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.example.authentication.CurrentUser;
 import com.example.controller.ApiResponse;
 import com.example.exception.CustomException;
 import com.example.exception.UnAuthorizedException;
 import com.example.exception.UserNotFoundException;
 import com.example.model.Address;
 import com.example.model.LoginDetails;
+import com.example.model.Role;
 import com.example.model.User;
+import com.example.model.UserToken;
 import com.example.repo.CartItemRepo;
 import com.example.repo.OrderRepo;
 import com.example.repo.UserRepo;
+import com.example.repo.UserTokenRepo;
 import jakarta.transaction.Transactional;
 
 @Service
 public class UserServiceImpl implements UserService{
 
-	
-	
 	private CartItemRepo cartItemRepo;
 	private OrderRepo orderRepo;
 	private UserRepo userRepo;
+	private CurrentUser currentUser;
+	private UserTokenRepo userTokenRepo;
 	
-	public UserServiceImpl(CartItemRepo cartItemRepo, OrderRepo orderRepo, UserRepo userRepo) {
+	public UserServiceImpl(CartItemRepo cartItemRepo, OrderRepo orderRepo, UserRepo userRepo, UserTokenRepo userTokenRepo, CurrentUser currentUser) {
 		this.cartItemRepo = cartItemRepo;
 		this.orderRepo = orderRepo;
 		this.userRepo = userRepo;
+		this.currentUser = currentUser;
+		this.userTokenRepo = userTokenRepo;
 		
 	}
 	public ResponseEntity<ApiResponse<User>> saveUser(User user) {
@@ -65,14 +73,20 @@ public class UserServiceImpl implements UserService{
 		if(!u.isPresent()) {
 			throw new UserNotFoundException("User Not Found");
 		}
+		
+		User currUser = currentUser.getUser();
+		if(currUser.getUserId() != userId) {
+			throw new UnAuthorizedException("Un Authorized");
+		}
+		
 		if(newUser.getUserName() == null) {
 			throw new UserNotFoundException("UserName Cannot be Empty");
 		}else if(newUser.getUserEmail() == null) {
 			throw new UserNotFoundException("UserEmail Cannot be Empty");
 		}else if(newUser.getUserPassword() == null) {
 			throw new UserNotFoundException("UserPassword Cannot be Empty");
-		}else if(newUser.getUserType() == null) {
-			throw new UserNotFoundException("UserType Cannot be Empty");
+		}else if(newUser.getUserRole() == null) {
+			throw new UserNotFoundException("UserRole Cannot be Empty");
 		}else if(newUser.getShippingAddress() == null) {
 			throw new UserNotFoundException("Shipping Address Cannot be Empty");
 		}else if(newUser.getPaymentDetails() == null) {
@@ -85,7 +99,7 @@ public class UserServiceImpl implements UserService{
 		oldUser.setUserPassword(newUser.getUserPassword());
 		oldUser.setShippingAddress(newUser.getShippingAddress());
 		oldUser.setPaymentDetails(newUser.getPaymentDetails());
-		oldUser.setUserType(newUser.getUserType());
+		oldUser.setUserRole(newUser.getUserRole());
 		
 		userRepo.save(oldUser);
 		ApiResponse<User> response = new ApiResponse<>();
@@ -96,11 +110,15 @@ public class UserServiceImpl implements UserService{
 
 	public ResponseEntity<ApiResponse<User>> getUserById(Long userId) {
 		Optional<User> exists = userRepo.findById(userId);
-		
 		if(!exists.isPresent()) {
-		
+			
 			throw new UserNotFoundException("User Not Found");
 		}
+		User currUser = currentUser.getUser();
+		if(currUser.getUserId()!= userId) {
+			throw new UnAuthorizedException("Not Authorized");
+		}
+		
 		User user = exists.get();
 		ApiResponse<User> response = new ApiResponse<>();
 		response.setData(user);
@@ -116,6 +134,11 @@ public class UserServiceImpl implements UserService{
 			
 			throw new UserNotFoundException("User Not Found");
 		}
+		User currUser = currentUser.getUser();
+		if(currUser.getUserId()!= userId) {
+			throw new UnAuthorizedException("Not Authorized");
+		}
+		
 		cartItemRepo.deleteAllByUser(userId);
 		orderRepo.deleteAllByUserId(userId);
 		userRepo.deleteById(userId);
@@ -126,27 +149,33 @@ public class UserServiceImpl implements UserService{
 		return ResponseEntity.ok(response);
 	}
 
-	public ResponseEntity<ApiResponse<List<User>>> getAllUsers() {
-		
-		List<User> list = userRepo.findAll();
-		ApiResponse<List<User>> response = new ApiResponse<>();
-		response.setData(list);
-		response.setMessage("All Users Details");
-		return ResponseEntity.ok(response);
-	}
+//	public ResponseEntity<ApiResponse<List<User>>> getAllUsers() {
+//		
+//		List<User> list = userRepo.findAll();
+//		ApiResponse<List<User>> response = new ApiResponse<>();
+//		response.setData(list);
+//		response.setMessage("All Users Details");
+//		return ResponseEntity.ok(response);
+//	}
 
 	public ResponseEntity<ApiResponse<User>> changeUserPassword(String eMail, String newPassword) {
-	    Optional<User> u = userRepo.findByUserEmail(eMail);
-
-	    if (!u.isPresent()) {
+		
+	    Optional<User> exists = userRepo.findByUserEmail(eMail);
+	    
+	    if (!exists.isPresent()) {
 	        throw new UnAuthorizedException("Invalid Email");
 	    }
 
 	    if (newPassword == null || newPassword.length()<=5) {
 	        throw new CustomException("New password cannot be empty or Less Than 5 Characters");
 	    }
-
-	    User user = u.get();
+	    
+	    User currUser = currentUser.getUser();
+		if(currUser.getUserId()!= exists.get().getUserId()) {
+			throw new UnAuthorizedException("Not Authorized");
+		}
+	    
+	    User user = exists.get();
 	    user.setUserPassword(newPassword);
 	    userRepo.save(user);
 	    ApiResponse<User> response = new ApiResponse<>();
@@ -164,15 +193,23 @@ public class UserServiceImpl implements UserService{
 		}
 		
 		User user = exists.get();
+		UserToken userToken = new UserToken();
 		
 		BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
 	    if (!encoder.matches(details.getLoginPassword(), user.getUserPassword())) {
 	        throw new UnAuthorizedException("Invalid credentials.");
 	    }
+	    String token = UUID.randomUUID().toString();
+	    userToken.setUserToken(token);
+	    userToken.setGeneratedAt(LocalDateTime.now());
+	    userToken.setUser(user);
+	    userRepo.save(user);
+	    userTokenRepo.save(userToken);
 	    
 		ApiResponse<User> response = new ApiResponse<>();
 		response.setMessage("Welcome User");
 		response.setData(user);
 		return ResponseEntity.ok(response);
 	}
+	
 }
